Added stuff for ia1

+++ NodeAddop.java
@@ -7,7 +7,7 @@ public class NodeAddop extends Node {
 	this.addop=addop;
     }
 
-    public int op(int o1, int o2) throws EvalException {
+    public Double op(double o1, double o2) throws EvalException {
 	if (addop.equals("+"))
 	    return o1+o2;
 	if (addop.equals("-"))
