+++ NodeAssn.java
@@ -8,7 +8,7 @@ public class NodeAssn extends Node {
 	this.expr=expr;
     }
 
-    public int eval(Environment env) throws EvalException {
+    public Double eval(Environment env) throws EvalException {
 	return env.put(id,expr.eval(env));
     }
