+++ NodeExpr.java
@@ -19,7 +19,7 @@ public class NodeExpr extends Node {
 	    this.expr.append(expr);
     }
 
-    public int eval(Environment env) throws EvalException {
+    public Double eval(Environment env) throws EvalException {
 	return expr==null
 	    ? term.eval(env)
 	    : addop.op(expr.eval(env),term.eval(env));
