+++ NodeFactNum.java
@@ -1,3 +1,7 @@
+//This is a subclass, of class NodeFact,
+//collectively modeling the parse-tree fact.
+//Takes in a string for a number
+//then is returned as parsed double.
 public class NodeFactNum extends NodeFact {
 
     private String num;
@@ -6,8 +10,8 @@ public class NodeFactNum extends NodeFact {
 	this.num=num;
     }
 
-    public int eval(Environment env) throws EvalException {
-	return Integer.parseInt(num);
+    public Double eval(Environment env) throws EvalException {
+	return Double.parseDouble(num);
     }
 
 }
